requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
idna==2.8
python-crontab==2.3.6
python-dateutil==2.8.0
requests==2.22.0
six==1.12.0
urllib3==1.25.3
